Rename permissions to runbook-permissions
