+++ backend/trpc/trpc.ts
@@ -1,7 +1,7 @@
 import { initTRPC, TRPCError } from "@trpc/server";
 import { ValidatedUser } from "backend/auth";
 import { Env } from "backend/types";
-import { PermissionsProvider } from "backend/permissions/types";
+import { PermissionsProvider } from "backend/runbook-permissions/types";
 
 export type TrcpContext = {
   env: Env;
